pyproject.toml
[project]
name = "voice_assistant"
version = "0.1.0"
description = "Voice assistant application"
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]

dependencies = [
    "requests>=2.31.0,<3.0.0",
    "rich>=13.7.1,<14.0.0",
    "torch>=2.2.1,<3.0.0",
    "openai-whisper @ git+https://github.com/openai/whisper.git",
    "sounddevice>=0.4.6,<0.5.0",
    "suno-bark @ git+https://github.com/suno-ai/bark.git",
    "speechrecognition>=3.10.1,<4.0.0",
    "pyaudio>=0.2.14,<0.3.0",
    "langchain>=0.2.0",
    "nltk>=3.8.1,<4.0.0",
    "langchain-chroma>=0.2.4,<0.3.0",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langchain-ollama (>=0.3.3,<0.4.0)"
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.6.2,<4.0.0"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry.group.dev.dependencies]
pylint = "^3.3.7"
isort = "^6.0.1"

